+++ packages/react/src/components/SummaryStats/src/components/SummaryStatsItem/index.ts
@@ -1,0 +1,5 @@
+import { SummaryStatsItem } from "./SummaryStatsItem";
+
+export { SummaryStatsItem } from "./SummaryStatsItem";
+export type { SummaryStatsItemProps } from "./summaryStatsItem.types";
+export default SummaryStatsItem;