+++ packages/react/src/components/SummaryStats/src/components/SummaryStatsStat/index.ts
@@ -1,5 +1,0 @@
-import { SummaryStatsStat } from "./SummaryStatsStat";
-
-export { SummaryStatsStat } from "./SummaryStatsStat";
-export type { SummaryStatsStatProps } from "./summaryStatsStat.types";
-export default SummaryStatsStat;