+++ packages/react/src/components/SummaryStats/src/components/index.ts
@@ -1,2 +1,2 @@
-export * from "./SummaryStatsStat";
+export * from "./SummaryStatsItem";
 export * from "./SummaryStatsTrendIndicator";